Keep updating tests in `tests/v1/05_backend/`
