+++ tests/v1/05_backend/test_backend_runners.py
@@ -13,6 +13,8 @@ Zurich.
 """
 import logging
 import os
+import sys
+from pathlib import Path
 
 import pytest
 from devtools import debug
@@ -21,9 +23,15 @@ from fractal_server.app.models.v1 import Workflow
 from fractal_server.app.routes.api.v1._aux_functions import (
     _workflow_insert_task,
 )
+from fractal_server.app.runner.executors.slurm._subprocess_run_as_user import (
+    _mkdir_as_user,
+)
 from fractal_server.app.runner.v1 import _backends
 from fractal_server.app.runner.v1.common import close_job_logger
 from fractal_server.logger import set_logger
+
+sys.path.append(Path(__file__).parent)
+from aux_create_subfolder import _create_task_subfolder  # noqa: E402
 
 
 def _extract_job_id_from_filename(filenames, pre, post) -> int:
@@ -113,16 +121,34 @@ async def test_runner(
     debug(wf)
 
     # Create working folder(s)
+    subfolder_names = []
     if backend == "local":
         workflow_dir = tmp777_path / "server"  # OK 777 here
         workflow_dir_user = workflow_dir
         umask = os.umask(0)
         workflow_dir.mkdir(parents=True, mode=0o700)
+        for wftask in wf.task_list:
+            subfolder = _create_task_subfolder(
+                wftask=wftask, workflow_dir=workflow_dir
+            )
+            subfolder_names.append(subfolder.name)
         os.umask(umask)
     elif backend == "slurm":
         workflow_dir, workflow_dir_user = request.getfixturevalue(
             "slurm_working_folders"
         )  # noqa
+        user = request.getfixturevalue("monkey_slurm_user")
+
+        umask = os.umask(0)
+        for wftask in wf.task_list:
+            subfolder = _create_task_subfolder(
+                wftask=wftask, workflow_dir=workflow_dir
+            )
+            subfolder_names.append(subfolder.name)
+            _mkdir_as_user(
+                folder=str(workflow_dir_user / subfolder.name), user=user
+            )
+        os.umask(umask)
 
     # Prepare backend-specific arguments
     logger_name = f"job_logger_{backend}"
@@ -168,8 +194,13 @@ async def test_runner(
     assert event2["parallelization"]["component_list"] == ["0", "1", "2"]
 
     # Check that the correct files are present in workflow_dir
-    files_server = [f.name for f in workflow_dir.glob("*")]
-    files_user = [f.name for f in workflow_dir_user.glob("*")]
+    files_server = [
+        f.name
+        for f in workflow_dir.glob(
+            "**/*",
+        )
+    ]
+    files_user = [f.name for f in workflow_dir_user.glob("**/*")]
     debug(sorted(files_server))
     debug(sorted(files_user))
 
@@ -182,11 +213,14 @@ async def test_runner(
     assert "2_par_0.err" in files_server
     assert "2_par_0.out" in files_server
     assert "2_par_0.metadiff.json" in files_server
-    with (workflow_dir_user / "0.args.json").open("r") as f:
-        debug(workflow_dir_user / "0.args.json")
+    with (workflow_dir_user / subfolder_names[0] / "0.args.json").open(
+        "r"
+    ) as f:
+        debug(workflow_dir_user / subfolder_names[0] / "0.args.json")
         args = f.read()
         debug(args)
         assert "logger_name" not in args
+
     # Check some backend-specific files
     # NOTE: the logic to retrieve the job ID is not the most elegant, but
     # it works (both for when a single or two SLURM backends are tested)
